Flips the layout of the mouse layer
The previous layout had the mouse keys in a non-standard arrangement.
This commit puts them into a logical place, based on the arrow
keys from the previous layer.

Switching between layers is somewhat confusing with this setup,
and should be investigated further. Long-term, a better solution may
account for switching between any number of layers.

Closes #30